requirements.txt
yt-dlp
PyQt5
requests
sponsorblock.py
pyperclip
pyinstaller